requirements.txt
streamlit==1.34.0
tensorflow==2.11.0
pillow==10.2.0
opencv-python-headless==4.8.0.76
